+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Change Log
 under development
 -----------------
 - Add `Context.JSON` to sends JSON response.
+- Add `Context.String` to sends string response.
 
 v1.8.1 April 2, 2020
 --------------------